chore: IoAdapter 설정 통합
