+++ apps/server/package.json
@@ -27,11 +27,14 @@
 		"@nestjs/jwt": "^10.2.0",
 		"@nestjs/passport": "^10.0.3",
 		"@nestjs/platform-express": "^10.0.0",
+		"@nestjs/platform-socket.io": "^10.4.8",
 		"@nestjs/typeorm": "^10.0.2",
+		"@nestjs/websockets": "^10.4.8",
 		"bcrypt": "^5.1.1",
 		"class-transformer": "^0.5.1",
 		"class-validator": "^0.14.1",
 		"dotenv": "^16.4.5",
+		"jsonwebtoken": "^9.0.2",
 		"lexorank": "^1.0.5",
 		"mysql2": "^3.11.4",
 		"passport": "^0.7.0",
@@ -39,6 +42,7 @@
 		"reflect-metadata": "^0.2.0",
 		"rxjs": "^7.8.1",
 		"sharedb": "^5.1.1",
+		"socket.io": "^4.8.1",
 		"typeorm": "^0.3.20"
 	},
 	"devDependencies": {
